Improve test coverage for clang-tidy-external
The patch ensures consistent naming, to undo parts of the confusion between
"clang-tidy-external" and "source-test-clang-external".

This patch adds testing so that clang-tidy-external does not report clang-diagnostic errors,
which ought to be caught by the other clang-tidy jobs or during built.

We also improve reporting logic to ensure that only `mozilla-civet`-prefixed
diagnotics are reported and last but not least update the reported error message.