Support the 'tgdiff' task to display changes in CI
This task runs anytime task configuration has run and uploads a series
of diff artifacts. The intent here is to shine a light on which tasks
are being added / removed for a given revision. Often it is easy to
accidentally cause changes on other branches (e.g beta) without
realizing it. Now reviewbot will comment with links to these artifacts
when they exist.