Refactor 'build_link' feature into re-usable notices
Notice tasks may or may not want to display links, and even if
they do, substituting the url into a hardcoded string in 'report.py'
isn't the most versatile way of going about it.

Instead, these tasks should just return a string and do their own url
substitution (if any). The reporter can then append each notice to the
overall comment.

This also removes the assumption that the 'docupload' task is the only
one using this feature.