Stand up a `TaskgraphDiffTask` to display task differences on Phabricator (#982)
* Split a 'BaseTask' object out of 'AnalysisTask'

* Implement a new 'NoticeTask' type and convert 'doc-upload' to use it

* Refactor 'build_link' feature into re-usable notices

Notice tasks may or may not want to display links, and even if
they do, substituting the url into a hardcoded string in 'report.py'
isn't the most versatile way of going about it.

Instead, these tasks should just return a string and do their own url
substitution (if any). The reporter can then append each notice to the
overall comment.

This also removes the assumption that the 'docupload' task is the only
one using this feature.

* Support the 'tgdiff' task to display changes in CI

This task runs anytime task configuration has run and uploads a series
of diff artifacts. The intent here is to shine a light on which tasks
are being added / removed for a given revision. Often it is easy to
accidentally cause changes on other branches (e.g beta) without
realizing it. Now reviewbot will comment with links to these artifacts
when they exist.

* Make some minor adjustments to 'docupload' message

This adds the 'NOTE' banner which makes it consistent with tgdiff as
well as more obvious that it is informational and not an issue. It also
does a bit of re-wording for clarity.

* Move 'build_comment' into the Phabricator reporter

The Phabricator reporter is the only thing using this function, and some
of the tasks contain Phabricator-flavoured markdown. So move this off
the base class for clarity.

Co-authored-by: ANdi <abpostelnicu@me.com>