restrict on push builds to `master` and `production` in taskcluster
In https://bugzilla.mozilla.org/show_bug.cgi?id=1907217 we're making scopes more explicit for branches. Because of this, pushes to branches not named in that repository will not be able to run. If this will be problematic, or we need to add more branches to this list, please let me know and we'll sort it out!